Deployment/requirements.txt
streamlit
pandas
matplotlib
opencv-python
squarify
tensorflow==2.15.0
Pillow